simulatori/requirements.txt
paho-mqtt>=2.0.0,<3.0.0
influxdb-client>=1.40.0,<2.0.0
